+++ content_restrictions/__init__.py
@@ -4,4 +4,4 @@ Init for the XblockContentRestrictions package.
 
 from .content_restrictions import XblockContentRestrictions
 
-__version__ = '0.1.0'
+__version__ = '0.2.0'