+++ content_restrictions/static/js/src/content_restrictions.js
@@ -3,6 +3,16 @@ function XblockContentRestrictions(runtime, element) {
 
     var handlerUrl = runtime.handlerUrl(element, 'has_access_with_password');
 
+    let gettext;
+    if ("ContentRestrictionsI18N" in window || "gettext" in window) {
+      gettext = window.ContentRestrictionsI18N?.gettext || window.gettext;
+    }
+
+    if (typeof gettext == "undefined") {
+      // No translations -- used by test environment
+      gettext = (string) => string;
+    }
+
     $(element).find('.submit-button').bind('click', function() {
         var data = {
             "password": $(element).find('input[name=password]').val(),
@@ -12,7 +22,7 @@ function XblockContentRestrictions(runtime, element) {
                 window.location.reload(false);
             }
             else {
-                $(element).find('.incorrect_password_explanation_text').text(response.error_message);
+                $(element).find('.incorrect_password_explanation_text').text(gettext(response.error_message));
             }
         }
         );